Refactor ConfiguratorIntegrationPanel
Refactor ConfiguratorIntegrationPanel.tsx into smaller components to improve maintainability.
